# contract/openapi.yaml
openapi: 3.0.3
info:
  title: AIDetector API
  version: 1.0.0
  description: V1.0 active contract baseline for auth, detection, history, quota, admin, report, key, and team APIs.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8000
    description: Local
...
paths:
...
  /api/v1/detect:
    post:
      tags: [detection]
      summary: Submit detection task
      operationId: detectText
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetectRequest'
      responses:
        '200':
          description: Detection accepted and persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Daily quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Downstream detection service failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/scan/examples:
    get:
      tags: [scan]
      summary: Get scan examples
      operationId: getScanExamples
      parameters:
        - name: locale
          in: query
          required: false
          schema:
            type: string
            default: zh-CN
      responses:
        '200':
          description: Scan examples
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanExamplesResponse'
  /api/v1/detections/parse-files:
    post:
      tags: [detection]
      summary: Parse uploaded files
      description: >
        Member-only legacy parser. This endpoint currently extracts plain text content only.
        It does not return structure-preserving HTML or block metadata.
      operationId: parseFiles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  maxItems: 5
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Parsed upload result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseFilesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/history:
    get:
      tags: [history]
      summary: List current user history
      operationId: listHistory
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: History list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [history]
      summary: Create history record
      operationId: createHistory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoryRecordCreateRequest'
      responses:
        '201':
          description: History record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryRecordResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [history]
      summary: Clear all current user history
      operationId: clearHistory
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All history cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClearAllResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/detections:
    get:
      tags: [admin]
      summary: List detections for admin
      operationId: listAdminDetections
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: actorType
          in: query
          schema:
            type: string
            enum: [user, guest]
        - name: label
          in: query
          schema:
            type: string
            enum: [ai, mixed, human]
        - name: function
          in: query
          schema:
            $ref: '#/components/schemas/DetectionFunction'
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, score, charsUsed]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Detection list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDetectionListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/detections/{detectionId}:
    get:
      tags: [admin]
      summary: Get detection detail for admin
      operationId: getAdminDetection
      security:
        - BearerAuth: []
      parameters:
        - name: detectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detection detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDetectionDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Detection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [admin]
      summary: Delete detection record
      operationId: deleteAdminDetection
      security:
        - BearerAuth: []
      parameters:
        - name: detectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Detection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - detail
      properties:
        code:
          oneOf:
            - type: integer
            - type: string
        message:
          type: string
        detail:
          nullable: true
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: string
...
    DetectionFunction:
      type: string
      enum: [scan]
...
    ScanExamplesResponse:
      type: object
      required:
        - locale
        - heroExamples
        - usageExamples
      properties:
        locale:
          type: string
        heroExamples:
          type: array
          items:
            $ref: '#/components/schemas/ScanHeroExampleItem'
        usageExamples:
          type: array
          items:
            $ref: '#/components/schemas/ScanUsageExampleItem'
...
    ParseFilesResponse:
      type: object
      description: Legacy plain-text file parsing response.
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ParsedFileResult'
...
    DetectRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
        functions:
          type: array
          items:
            $ref: '#/components/schemas/DetectionFunction'
          default: [scan]
        options:
          type: object
          additionalProperties: true
          nullable: true
        editorHtml:
          type: string
          nullable: true
    DetectResponse:
      type: object
      required:
        - detectionId
        - historyId
        - label
        - score
        - currentCredits
        - inputText
      properties:
        detectionId:
          type: integer
        historyId:
          type: integer
        label:
          type: string
          enum: [ai, mixed, human]
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        modelName:
          type: string
          nullable: true
        rawScore:
          type: number
          format: float
          nullable: true
        threshold:
          type: number
          format: float
          nullable: true
        currentCredits:
          type: integer
        inputText:
          type: string
        result:
          $ref: '#/components/schemas/HistoryAnalysis'
          nullable: true
...
    HistoryRecordCreateRequest:
      type: object
      required:
        - title
        - functions
        - inputText
        - editorHtml
      properties:
        title:
          type: string
          maxLength: 255
        functions:
          type: array
          items:
            $ref: '#/components/schemas/DetectionFunction'
        inputText:
          type: string
          maxLength: 50000
        editorHtml:
          type: string
        analysis:
          $ref: '#/components/schemas/HistoryAnalysis'
          nullable: true
...
    HistoryListResponse:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/HistoryRecordResponse'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    ClearAllResponse:
      type: object
      required:
        - deletedCount
      properties:
        deletedCount:
          type: integer
...